+++ src/datasources/results/query-handlers/query-results/QueryResultsDataSource.test.ts
@@ -80,6 +80,25 @@ describe('QueryResultsDataSource', () => {
       expect(response.data).toMatchSnapshot();
     });
 
+    test('should set the default order by to "STARTED_AT" and descending to "true"', async () => {
+      const query = buildQuery({
+        refId: 'A',
+        outputType: OutputType.Data
+      });
+
+      await datastore.query(query);
+
+      expect(backendServer.fetch).toHaveBeenCalledTimes(1);
+      expect(backendServer.fetch).toHaveBeenCalledWith(
+        expect.objectContaining({
+          data: expect.objectContaining({
+            orderBy: 'STARTED_AT',
+            descending: true,
+          }),
+        })
+      )
+    });
+
     test('returns total count for valid total count output type queries', async () => {
       const query = buildQuery({
         refId: 'A',
@@ -90,6 +109,25 @@ describe('QueryResultsDataSource', () => {
 
       expect(response.data).toHaveLength(1);
       expect(response.data).toMatchSnapshot();
+    });
+
+    test('returns totalCount as 0 in query() when OutputType is TotalCount and no results', async () => {
+      backendServer.fetch
+        .calledWith(requestMatching({ url: '/nitestmonitor/v2/query-results', method: 'POST' }))
+        .mockReturnValue(createFetchResponse({
+          results: [],
+          totalCount: 0
+        } as QueryResultsResponse));
+
+      const query = buildQuery({
+        refId: 'A',
+        outputType: OutputType.TotalCount
+      });
+
+      const response = await datastore.query(query);
+
+      expect(response.data[0].fields).toEqual([{ name: 'Total count', values: [0] }]);
+      expect(response.data[0].refId).toBe('A');
     });
 
     test('returns no data when QueryResults API returns empty array', async () => {
@@ -584,6 +622,22 @@ describe('QueryResultsDataSource', () => {
 
       expect(result).toEqual([]);
     });
+
+    test('should set the default order by to "started at" and descending to true', async () =>{
+      const query = { properties: ResultsPropertiesOptions.PART_NUMBER, queryBy: '', resultsTake: 1000 } as ResultsVariableQuery;
+
+      await datastore.metricFindQuery(query, {});
+
+      expect(backendServer.fetch).toHaveBeenCalledTimes(1);
+      expect(backendServer.fetch).toHaveBeenCalledWith(
+        expect.objectContaining({
+          data: expect.objectContaining({
+            orderBy: 'STARTED_AT',
+            descending: true,
+          }),
+        })
+      );
+    })
 
      it.each([-1, NaN, 10001])('should return empty array if resultsTake value is invalid (%p)', async (invalidResultsTake) => {
             const query = {