+++ src/datasources/results/query-handlers/query-steps/QueryStepsDataSource.test.ts
@@ -113,6 +113,25 @@ describe('QueryStepsDataSource', () => {
       expect(response.data).toMatchSnapshot();
     });
 
+    test('should set default orderby to "STARTED_AT" and descending to "false"', async () => {
+      const query = buildQuery({
+        refId: 'A',
+        outputType: OutputType.Data,
+      });
+
+      await datastore.query(query);
+
+      expect(backendServer.fetch).toHaveBeenCalledWith(
+        expect.objectContaining({
+          url: '/nitestmonitor/v2/query-steps',
+          data: expect.objectContaining({
+            orderBy: "STARTED_AT",
+            descending: false,
+          }),
+        })
+      );
+    });
+
     test('should return total count for valid total count output type queries', async () => {
       const query = buildQuery({
         refId: 'A',
@@ -1192,6 +1211,22 @@ describe('QueryStepsDataSource', () => {
         const result = await datastore.metricFindQuery(query);
 
         expect(result).toEqual([]);
+      });
+
+      it('should set default orderby to "STARTED_AT" and descending to "false"', async () => {
+        const query = { queryByResults: 'programName = "name"', stepsTake: 1000, partNumberQueryInSteps: ['PartNumber1'] } as StepsVariableQuery;
+
+        await datastore.metricFindQuery(query);
+
+        expect(backendServer.fetch).toHaveBeenCalledWith(
+          expect.objectContaining({
+            url: '/nitestmonitor/v2/query-steps',
+            data: expect.objectContaining({
+              orderBy: "STARTED_AT",
+              descending: false,
+            }),
+          })
+        );
       });
 
       it.each([-1, NaN, 10001])('should return empty array if stepsTake value is invalid (%p)', async (invalidStepsTake) => {