+++ src/datasources/work-orders/constants/WorkOrdersQueryBuilder.constants.ts
@@ -147,10 +147,6 @@ export const WorkOrdersQueryBuilderStaticFields = [
   WorkOrdersQueryBuilderFields.WORK_ORDER_ID,
   WorkOrdersQueryBuilderFields.STATE,
   WorkOrdersQueryBuilderFields.TYPE,
-  WorkOrdersQueryBuilderFields.ASSIGNED_TO,
-  WorkOrdersQueryBuilderFields.REQUESTED_BY,
-  WorkOrdersQueryBuilderFields.CREATED_BY,
-  WorkOrdersQueryBuilderFields.UPDATED_BY,
   WorkOrdersQueryBuilderFields.PROPERTIES,
   WorkOrdersQueryBuilderFields.DUE_DATE,
 ];